feat: store posts tx hashes (#157)
## Description

This PR adds the ability to store the hashes of transactions related to
each post. This is particularly useful for clients that need to sync
posts created locally once they are sent on the chain.

To do this we create a new `post_tx_hash` table inside the database, and
each time a `MsgCreatePost`, `MsgEditPost`, `MsgAddPostAttachment` or
`MsgRemovePostAttachment` is broadcasted, we store inside such table the
hash of the transaction as well as the reference to the post. This
allows to expose a new `transactions` field inside the GraphQL `post`
type, containing a list of all the transactions related to that post.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

I have...

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] added `!` to the type prefix if API or client breaking change
- [x] targeted the correct branch
- [ ] provided a link to the relevant issue or specification
- [ ] added a changelog entry to `CHANGELOG.md`
- [x] included comments for [documenting Go
code](https://blog.golang.org/godoc)
- [ ] updated the relevant documentation or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

I have...

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed `!` in the type prefix if API or client breaking change
- [ ] confirmed all author checklist items have been addressed
- [ ] reviewed API design and naming
- [ ] reviewed documentation is accurate
- [ ] reviewed tests and test coverage
- [ ] manually tested (if applicable)